moved reply to journalist_app/main.py
