+++ securedrop/journalist.py
@@ -9,7 +9,7 @@ import config
 import crypto_util
 from flask_babel import gettext
 import store
-from db import db_session, Source, Submission, Reply
+from db import db_session, Source, Submission
 
 from journalist_app import create_app
 from journalist_app.decorators import login_required
@@ -99,58 +99,6 @@ def download_single_submission(filesystem_id, fn):
 
     return send_file(store.path(filesystem_id, fn),
                      mimetype="application/pgp-encrypted")
-
-
-@app.route('/reply', methods=('POST',))
-@login_required
-def reply():
-    """Attempt to send a Reply from a Journalist to a Source. Empty
-    messages are rejected, and an informative error message is flashed
-    on the client. In the case of unexpected errors involving database
-    transactions (potentially caused by racing request threads that
-    modify the same the database object) logging is done in such a way
-    so as not to write potentially sensitive information to disk, and a
-    generic error message is flashed on the client.
-
-    Returns:
-       flask.Response: The user is redirected to the same Source
-           collection view, regardless if the Reply is created
-           successfully.
-    """
-    form = ReplyForm()
-    if not form.validate_on_submit():
-        for error in form.message.errors:
-            flash(error, "error")
-        return redirect(url_for('col', filesystem_id=g.filesystem_id))
-
-    g.source.interaction_count += 1
-    filename = "{0}-{1}-reply.gpg".format(g.source.interaction_count,
-                                          g.source.journalist_filename)
-    crypto_util.encrypt(form.message.data,
-                        [crypto_util.getkey(g.filesystem_id),
-                         config.JOURNALIST_KEY],
-                        output=store.path(g.filesystem_id, filename))
-    reply = Reply(g.user, g.source, filename)
-
-    try:
-        db_session.add(reply)
-        db_session.commit()
-    except Exception as exc:
-        flash(gettext(
-            "An unexpected error occurred! Please "
-            "inform your administrator."), "error")
-        # We take a cautious approach to logging here because we're dealing
-        # with responses to sources. It's possible the exception message could
-        # contain information we don't want to write to disk.
-        app.logger.error(
-            "Reply from '{}' (ID {}) failed: {}!".format(g.user.username,
-                                                         g.user.id,
-                                                         exc.__class__))
-    else:
-        flash(gettext("Thanks. Your reply has been stored."),
-              "notification")
-    finally:
-        return redirect(url_for('col', filesystem_id=g.filesystem_id))
 
 
 @app.route('/regenerate-code', methods=('POST',))
