+++ securedrop/journalist_templates/col.html
@@ -73,7 +73,7 @@
   <h3><i class="fa fa-reply"></i> {{ gettext('Reply') }}</h3>
   {% if source.has_key %}
     <p>{{ gettext('You can write a secure reply to the person who submitted these documents:') }}</p>
-    <form action="/reply" method="post" autocomplete="off">
+    <form action="{{ url_for('main.reply') }}" method="post" autocomplete="off">
       <input name="csrf_token" type="hidden" value="{{ csrf_token() }}">
       <input type="hidden" name="filesystem_id" value="{{ filesystem_id }}">
       {{ form.message(id="reply-text-field", rows='10', cols='72') }}
